Updated styles on summary page

+++ db/migrate/20150421133256_create_forms.rb
@@ -6,17 +6,17 @@ class CreateForms < ActiveRecord::Migration
       t.integer :income
       t.string :bedrooms
       t.string :healthcare
-      t.decimal :savings
+      t.integer :savings
       t.integer :dining_out_low
       t.integer :dining_out_medium
       t.integer :dining_out_high
       t.integer :driving_trips
       t.integer :car_trip_duration
       t.integer :mass_transit_trips
-      t.decimal :groceries
-      t.decimal :cabs
-      t.decimal :recreation
-      t.decimal :shopping
+      t.integer :groceries
+      t.integer :cabs
+      t.integer :recreation
+      t.integer :shopping
       t.integer :random_path
       t.timestamps null: false
     end
